requirements.txt
fastapi
uvicorn
pillow
numpy
pycryptodome
qiskit
qiskit-aer
python-multipart

